test: 合并 pages 与 app_release 测试文件
- test_pages_guest_and_redirects 并入 test_pages.py（未登录重定向/登录页渲染）
- 删除 test_pages_full.py（4 例弱断言 status in [200,302,404] 被覆盖）
- test_app_release_api/prerelease 并入 test_app_release.py，删除 2 个与 HTTP 层重复的函数级用例
